Consultations : The "max voters per choice" option is enabled.
The group's filter is now used for consultations too. 

The right ADD_CONSULTATION works. 

Verify a user has the right to create/see/vote a consultation before creating/seeing/voting it in case of the user managed to access the creation/display/vote action. 

Note : All the options for the consultations are finished, please to test it for debug.

git-svn-id: http://fsnet.googlecode.com/svn@1901 07c314b0-eee8-11dd-b478-27b5dcd53177